+++ backend/src/main/java/com/project/harmonie_e_commerce/dto/ProductImageDTO.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @Builder
 public class ProductImageDTO {
     @JsonProperty("product_id")
-    private Long productId;
+    private Integer productId;
 
     @JsonProperty("image_url")
     private String imageUrl;
